matrix-server/requirements.txt
Adafruit-Blinka==5.8.0
adafruit-circuitpython-neopixel==6.0.0
adafruit-circuitpython-pypixelbuf==2.2.0
Adafruit-PlatformDetect==2.22.1
Adafruit-PureIO==1.1.7
art==5.0
asn1crypto==0.24.0
cbor2==5.2.0
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
cryptography==2.6.1
Deprecated==1.2.10
entrypoints==0.3
Flask==1.1.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==17.1.1
keyrings.alt==3.1.1
luma.core==2.0.1
luma.led-matrix==1.5.0
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
pycrypto==2.6.1
pyftdi==0.52.0
PyGObject==3.30.4
pyserial==3.4
python-apt==1.8.4.1
pyusb==1.1.0
pyxdg==0.25
requests==2.21.0
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
scipy==1.5.4
SecretStorage==2.3.1
six==1.12.0
smbus2==0.3.0
spidev==3.5
ssh-import-id==5.7
sysv-ipc==1.0.1
urllib3==1.24.1
virtualenv==15.1.0
Werkzeug==1.0.1
wrapt==1.12.1
ws2812==1.0.0
